+++ lib/dentaku/ast/case/case_switch_variable.rb
@@ -30,6 +30,12 @@ module Dentaku
       def accept(visitor)
         visitor.visit_switch(self)
       end
+
+      private
+
+      def compute_pure?
+        @node.pure?
+      end
     end
   end
 end